move models from const to models, display no data message
