+++ MediaSet.Remix/app/constants.ts
@@ -1,12 +1,2 @@
 // export const baseUrl = 'http://localhost:7130';
 export const baseUrl = process.env.apiUrl;
-
-export const entities = {
-  books: "Books",
-  movies: "Movies",
-};
-
-export enum Entity {
-  Books = "Books",
-  Movies = "Movies",
-};
