+++ MediaSet.Remix/app/routes/books2.$bookId_.delete/route.tsx
@@ -1,9 +1,8 @@
 import type { ActionFunctionArgs } from "@remix-run/node";
 import { redirect } from "@remix-run/node";
 import invariant from "tiny-invariant"
-import { Entity } from "~/constants";
 import { deleteEntity } from "~/entity-data";
-import { Book } from "~/models";
+import { Book, Entity } from "~/models";
 
 export const action = async ({ params }: ActionFunctionArgs) => {
   invariant(params.bookId, "Missing bookId param");
